# Conflicts:
#	gulpfile.js
#	src/layaAir/Laya.ts
#	src/layaAir/laya/RenderEngine/RenderInterface/RenderPipelineInterface/IRenderQueue.ts
#	src/layaAir/laya/RenderEngine/UniformBufferObject.ts
#	src/layaAir/laya/components/Script.ts
#	src/layaAir/laya/d3/RenderObjs/NativeOBJ/NativeBaseRenderQueue.ts
#	src/layaAir/laya/d3/RenderObjs/RenderObj/BaseRenderNode.ts
#	src/layaAir/laya/d3/RenderObjs/RenderObj/BaseRenderQueue.ts
#	src/layaAir/laya/d3/core/Camera.ts
#	src/layaAir/laya/d3/core/MeshRenderer.ts
#	src/layaAir/laya/d3/core/SimpleSkinnedMeshRenderer.ts
#	src/layaAir/laya/d3/core/scene/Scene3D.ts
#	src/layaAir/laya/d3/utils/Utils3D.ts
#	src/layaAir/laya/display/Node.ts
#	src/layaAir/laya/display/Sprite.ts
#	src/layaAir/laya/net/Loader.ts
#	src/layaAir/laya/physics/Physics2D.ts
#	src/layaAir/laya/resource/RenderTexture.ts
#	src/layaAir/laya/webgl/canvas/DrawStyle.ts

+++ src/layaAir/laya/Physics3D/Bullet/Shape/btMeshColliderShape.ts
@@ -92,7 +92,7 @@ export class btMeshColliderShape extends btColliderShape implements IMeshCollide
 		let posArray = new Array<Vector3>();
 		value.getPositions(posArray);
 
-		var indices: Uint16Array = value._indexBuffer.getData();//TODO:API修改问题
+		var indices: Uint16Array = value._indexBuffer.getData() as Uint16Array;//TODO:API修改问题
 		for (var i: number = 0, n: number = indices.length; i < n; i += 3) {
 			var position0: Vector3 = posArray[indices[i]];
 			var position1: Vector3 = posArray[indices[i + 1]];
